refactor: notice모달 리팩토링중
